Refactor w3tracecontext integration test (#3563)
